backend/requirements.txt
tensorflow==2.0.0b0
numpy
Pillow